fix: robustify pybullet simulation's contact check (#24)
In the `python3 scripts/pybullet_interactive_demo.py`, sometimes the
`getContactPoints` would return no points even though the distance
between the balls is less then the sum of their radii, which currently
causes the robot to abruptly stop the motors and the camera to pan to
the default view. Checking `getClosestPoints` instead / in addition
avoids this issue.